+++ examples/non_blocking_iid_example.py
@@ -1,3 +1,4 @@
+from time import sleep
 from decanter_ai_sdk.non_blocking_client import Client
 import os
 from decanter_ai_sdk.enums.evaluators import ClassificationMetric
@@ -11,7 +12,9 @@ def test_iid():
     host = ""  # TODO fill in real host
     print("---From test iid---")
 
-    client = Client(auth_key=auth_key, project_id=project_id, host=host)
+    client = Client(
+        auth_key=auth_key, project_id=project_id, host=host, dry_run_type=None
+    )
 
     current_path = os.path.dirname(os.path.abspath(__file__))
 
@@ -19,9 +22,22 @@ def test_iid():
     train_file = open(train_file_path, "rb")
     train_id = client.upload(train_file, "train_file")
 
+    while client.check_upload_status(train_id) != "done":
+        print(
+            "Upload task: ", train_id, " status: ", client.check_upload_status(train_id)
+        )
+        sleep(3)
+    print("Train file uploaded.")
+
     test_file_path = os.path.join(current_path, "../data/test.csv")
     test_file = open(test_file_path, "rb")
     test_id = client.upload(test_file, "test_file")
+    while client.check_upload_status(train_id) != "done":
+        print(
+            "Upload task: ", test_id, " status: ", client.check_upload_status(test_id)
+        )
+        sleep(3)
+    print("Test file uploaded.")
 
     print("This will show top 2 uploaded table names and ids: \n")
 
@@ -30,11 +46,11 @@ def test_iid():
         print(count, "name:", table["name"], ",id:", table["_id"])
 
     print(
-        "\nThis will show the info of the first table:",
+        "\nThis will show the info of the first table:\n",
         client.get_table(client.get_table_list()[0]["_id"]),
     )
 
-    experiment = client.train_iid(
+    experiment_id = client.train_iid(
         experiment_name="exp_name",
         experiment_table_id=train_id,
         target="Survived",
@@ -47,12 +63,21 @@ def test_iid():
         algos=["DRF", "GBM", IIDAlgorithms.DRF],
     )
 
-    print("This will show info of the experiment:\n", experiment.experiment_info())
+    # print("This will show info of the experiment:\n", experiment.experiment_info())
 
-    m = experiment.get_best_model()
+    while client.check_exp_status(experiment_id) != "done":
+        print("Experiment not done yet.")
+        sleep(3)
+
+    print("Experiment id:", experiment_id, " done.")
+
+    exp = client.get_exp_result(exp_id=experiment_id)
+
+    m = exp.result.get_best_model()
+
     print("This will show the best model id:", m.model_id, "name:", m.model_name, "\n")
 
-    model = experiment.get_best_model_by_metric(ClassificationMetric.MISCLASSIFICATION)
+    model = exp.result.get_best_model_by_metric(ClassificationMetric.MISCLASSIFICATION)
     print(
         "This will show the best model evaluated by misclassification. id:",
         model.model_id,
@@ -61,11 +86,25 @@ def test_iid():
         "\n",
     )
 
-    predict = client.predict_iid(
+    predict_id = client.predict_iid(
         keep_columns=[], non_negative=False, test_table_id=test_id, model=m
     )
+    while client.check_pred_status(predict_id) != "done":
+        print("Prediction task not done yet.")
+        sleep(3)
+
+    print("Prediction id:", predict_id, " done.")
+
+    prediction = client.get_pred_result(pred_id=predict_id)
 
     print(
-        "This will show the id of tested ../data:", predict.attributes["table_id"], "\n"
+        "This will show the id of tested ../data:",
+        prediction.result.attributes["table_id"],
+        "\n",
     )
-    print("Head of the prediction ../data:\n", predict.get_predict_df().head())
+    print(
+        "Head of the prediction ../data:\n", prediction.result.get_predict_df().head()
+    )
+
+
+test_iid()
